+++ cpp/src/db/DBImpl.cpp
@@ -198,18 +198,25 @@ Status DBImpl::Query(const std::string& table_id, const std::vector<std::string>
         ids.push_back(std::stoul(id, &sz));
     }
 
-    meta::TableFilesSchema files_array;
-    auto status = meta_ptr_->GetTableFiles(table_id, ids, files_array);
+    meta::DatePartionedTableFilesSchema files_array;
+    auto status = meta_ptr_->FilesToSearch(table_id, ids, dates, files_array);
     if (!status.ok()) {
         return status;
     }
 
-    if(files_array.empty()) {
+    meta::TableFilesSchema file_id_array;
+    for (auto &day_files : files_array) {
+        for (auto &file : day_files.second) {
+            file_id_array.push_back(file);
+        }
+    }
+
+    if(file_id_array.empty()) {
         return Status::Error("Invalid file id");
     }
 
     cache::CpuCacheMgr::GetInstance()->PrintInfo(); //print cache info before query
-    status = QueryAsync(table_id, files_array, k, nq, vectors, dates, results);
+    status = QueryAsync(table_id, file_id_array, k, nq, vectors, dates, results);
     cache::CpuCacheMgr::GetInstance()->PrintInfo(); //print cache info after query
     return status;
 }
@@ -357,7 +364,7 @@ void DBImpl::StartCompactionTask() {
 
 Status DBImpl::MergeFiles(const std::string& table_id, const meta::DateT& date,
         const meta::TableFilesSchema& files) {
-    ENGINE_LOG_DEBUG << "Merge files for table" << table_id;
+    ENGINE_LOG_DEBUG << "Merge files for table " << table_id;
 
     meta::TableFileSchema table_file;
     table_file.table_id_ = table_id;
@@ -424,13 +431,15 @@ Status DBImpl::BackgroundMergeFiles(const std::string& table_id) {
     bool has_merge = false;
     for (auto& kv : raw_files) {
         auto files = kv.second;
-        if (files.size() <= options_.merge_trigger_number) {
+        if (files.size() < options_.merge_trigger_number) {
+            ENGINE_LOG_DEBUG << "Files number not greater equal than merge trigger number, skip merge action";
             continue;
         }
         has_merge = true;
         MergeFiles(table_id, kv.first, kv.second);
 
         if (shutting_down_.load(std::memory_order_acquire)){
+            ENGINE_LOG_DEBUG << "Server will shutdown, skip merge action for table " << table_id;
             break;
         }
     }
@@ -447,6 +456,11 @@ void DBImpl::BackgroundCompaction(std::set<std::string> table_ids) {
         if (!status.ok()) {
             ENGINE_LOG_ERROR << "Merge files for table " << table_id << " failed: " << status.ToString();
             continue;//let other table get chance to merge
+        }
+
+        if (shutting_down_.load(std::memory_order_acquire)){
+            ENGINE_LOG_DEBUG << "Server will shutdown, skip merge action";
+            break;
         }
     }
 
@@ -581,6 +595,11 @@ Status DBImpl::BuildIndexByTable(const std::string& table_id) {
             return status;
         }
         ENGINE_LOG_DEBUG << "Sync building index for " << file.id_ << " passed";
+
+        if (shutting_down_.load(std::memory_order_acquire)){
+            ENGINE_LOG_DEBUG << "Server will shutdown, skip build index action for table " << table_id;
+            break;
+        }
     }
 
     return status;
@@ -601,6 +620,7 @@ void DBImpl::BackgroundBuildIndex() {
         }
 
         if (shutting_down_.load(std::memory_order_acquire)){
+            ENGINE_LOG_DEBUG << "Server will shutdown, skip build index action";
             break;
         }
     }